Release Candidate 2 `Flawless`
Rtp and Rtcp Packet were completely re-written to use IDisposable and support Packet completion.

RtpHeader and RtcpHeader have been seperated from the packet entities.

CommonHeaderBits supplies a base class to unify those headers under.

IPacket interface was defined to consolidate Rtsp, Rtp and Rtcp packets.

Binary class was added to faciliate bit manipulation in non system endian.

A SourceList class was added to represent the SourceLists abstraction present in RtpPackets and certain RtcpPackets.

RtpExtension was added to faciliate the reading of RtpExtesnion headers and the data contained in the RtpExtension.

A new mechanism for mapping Rtcp implemeations was provided.

RtcpPackets  implementations which did not belong to RFC3550 were removed for now.

All Rtcp Implementions were updated to reflect the changes in RtcpPacket.

RtcpReport was made abstract and all reports now inherit from RtcpReport.

ReportBlock was modified to be a fixed size contigous allocation of memory.

IReportBlock interface was added to consolidate all and unify all further implementations of ReportBlocks.

SourceDescription and SourceDescriptionItem has been updated to be complaint with RFC3550 and the changes in ReportBlock.

SourceDescriptionChunk was changed added to faciliate the changes in ReportBlock and SourceDescription.

Changes to the RtpClient to faciliate Multicast sending and receiving as well the other listed changes.

Conference class was added to faciliate using a RtpClient with more then 1 party.

RtpFrame was updated to be complaint with RFC3550 and further implemetnations.

JpegFrame was renamed to RFC2435 frame and does not allow RFC2035 formatted packets by default.

RtpDump classes were updated to facilicate these changes.

RtspClient was updated to reflect these changes.

+more, check out the diff!