Added JpegSouceStream, MJpegSourceStream classes for Http based sources to allow for sending them out via Rtp.
Updated SourceList to have IReadOnlyCollection

Updated SourceDescriptionItemList to have IReadOnlyCollection

Added RtmpPacket.

Added basis for RtmpClient

Added basis for Rtcp Feedback

Consolidated ISource and ISink with IMediaStream but left them for the ability to sort through types and for the ability to have a dual implementation. Renamed to IMediaSource and IMediaSink.