MediaFileStream - Write, WriteByte, BeginWrite and WriteAsync have been overridden and AbsoluteRead, AbsoluteWrite and Append has been added.
This will allow use of the readers when reading from a remote file source as well as when writing to the same file.

Node - RawData is now Data because RawData would imply all data.

Data and UpdateData now utilize the MediaFileStream AbsoluteRead and AbsoluteWrite methods if the IMediaContainer.BaseStream is a MedaFileStream for better performance.