*Stable Nightly Container Release*
RtpDumpReader - Added change to make reader compatible with bark.rtp however it is not compatible with the files written by RtpDumpWriter / Goole and the Spec anymore. This will be addressed with either an option or a detection on the data to determine which format it is.

RtpToolEntry - Start, Source and Port are set to definite values from the top of the structure, Pointer only controls the offset to which the RD_packet_t begins.

RtcpPacket - PayloadType not retrieved when parsing packets unless given.

Added comments explaining how packets are parsed.

Cleaned up GetPackets.