- This is the last release scheduled for 'Container' -
[Bug Fixes, Compliance, Performance]

RtpClient - IsConnected and Connect have been renamed to IsActive and Activate.

MemorySegment. Added Empty.

Binary.ReadBitsReverse was revised to work in unaligned cases.

Concepts was updated with the classes which will evevntually be used on the RtspClient and RtpClient to support configuration and connection pooling.

RtpClient no longer calls SetBufferSize when connecting.

RtpFrame now allows multiple payload types to be contained.

Adds default constructor for RtpFrame which can allow settings of the PayloadType when the first packet is added.

RFC3640Frame - Fixed bugs with reading padded values.