*Formats Beta*
If I were to dissertate the subtle details on all of the changes included it would be a complete work of it's own.

Please provide feedback, without feedback I will not evolve this library or fix subtle bugs nor provide support.

I hope that you will find this release fairly stable and a good indication of things to come if such feedback is provided.

Please remember I take my free time and energy to develop and engineer this library and I have plenty of other things that I probably could be doing...

|.v//.|